Refactor exitStatus structure
